ai_analyst_mvp/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
sentence-transformers>=2.2.2
chromadb
rank_bm25
numpy
scikit-learn
tqdm
python-dotenv
requests
transformers
faiss-cpu
accelerate
pandas
nltk
openai>=1.42.0


